- separate main and sample classes - update screenshots
