+++ cf/samples/java-samples/commons/src/main/java/commons/SampleException.java
@@ -1,0 +1,24 @@
+package commons;
+
+public class SampleException
+extends Exception {
+
+	private static final long serialVersionUID = 1L;
+
+	public SampleException() {
+		super();
+	}
+
+	public SampleException(String message) {
+		super(message);
+	}
+
+	public SampleException(String message, Throwable cause) {
+		super(message, cause);
+	}
+
+	public SampleException(Throwable cause) {
+		super(cause);
+	}
+
+}
